+++ packages/components/src/lab/field/Field.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import '@testing-library/jest-dom';
-import { render, screen, cleanup } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
+import { render, screen, cleanup } from '../../test-utils';
 import Field from '.';
 
 let props;