+++ packages/components/src/lab/pagination/paginationLink/PaginationLink.js
@@ -1,8 +1,13 @@
 import React from 'react';
 import Types from 'prop-types';
 import classNames from 'classnames';
+import { useIntl } from 'react-intl';
+
+import messages from './PaginationLink.messages';
 
 const PaginationLink = ({ disabled, active, pageNumber, onClick, children }) => {
+  const intl = useIntl();
+
   const handleClick = (e) => {
     e.preventDefault();
     if (!disabled && !active) {
@@ -16,8 +21,10 @@ const PaginationLink = ({ disabled, active, pageNumber, onClick, children }) => 
     }
   };
 
-  const ariaLabel = active ? `Go to page ${pageNumber}` : `Current page, page ${pageNumber}`;
-
+  const ariaLabel = intl.formatMessage(
+    active ? messages.ariaLabelActive : messages.ariaLabelInactive,
+    { pageNumber },
+  );
   return (
     <li>
       <a