+++ packages/components/src/lab/pagination/paginationLink/PaginationLink.messages.js
@@ -1,0 +1,12 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  ariaLabelActive: {
+    id: 'neptune.PaginationLink.ariaLabel.active',
+    defaultMessage: 'Go to page {pageNumber}',
+  },
+  ariaLabelInactive: {
+    id: 'neptune.PaginationLink.ariaLabel.inactive',
+    defaultMessage: 'Current page, page {pageNumber}',
+  },
+});