+++ packages/components/src/navigationOption/NavigationOption.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { shallow } from 'enzyme';
-import { render } from '@testing-library/react';
+import { render } from '../test-utils';
 
 import NavigationOption from '.';
 import Option from '../common/Option';