+++ packages/components/src/overlayHeader/OverlayHeader.spec.js
@@ -1,6 +1,5 @@
 import React from 'react';
-import '@testing-library/jest-dom';
-import { render } from '@testing-library/react';
+import { render } from '../test-utils';
 
 import OverlayHeader from '.';
 import Logo from '../common/logo';