+++ packages/docs/utils/sections.js
@@ -74,6 +74,10 @@ export default [
         dir: 'components/overlays',
       },
       {
+        title: 'Context',
+        dir: 'components/context',
+      },
+      {
         title: 'Dynamic Forms',
         dir: 'components/dynamicForms',
       },