+++ packages/docs/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.3.1](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@8.3.0...@transferwise/neptune-docs@8.3.1) (2020-12-16)
+
+**Note:** Version bump only for package @transferwise/neptune-docs
+
+
+
+
+
+# [8.3.0](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@8.2.4...@transferwise/neptune-docs@8.3.0) (2020-12-16)
+
+
+### Features
+
+* internalization support ([#674](https://github.com/transferwise/neptune-web/issues/674)) ([10c533f](https://github.com/transferwise/neptune-web/commit/10c533fd9af2bb25d0f9b195ea941989252bf09e))
+
+
+
+
+
 ## [8.2.4](https://github.com/transferwise/neptune-web/compare/@transferwise/neptune-docs@8.2.3...@transferwise/neptune-docs@8.2.4) (2020-12-14)