+++ packages/dynamic-flows/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [8.0.3](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@8.0.2...@transferwise/dynamic-flows@8.0.3) (2020-12-16)
+
+**Note:** Version bump only for package @transferwise/dynamic-flows
+
+
+
+
+
+## [8.0.2](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@8.0.1...@transferwise/dynamic-flows@8.0.2) (2020-12-16)
+
+**Note:** Version bump only for package @transferwise/dynamic-flows
+
+
+
+
+
 ## [8.0.1](https://github.com/transferwise/neptune-web/compare/@transferwise/dynamic-flows@8.0.0...@transferwise/dynamic-flows@8.0.1) (2020-12-14)